Crate eq_bailsman

Equilibrium Bailsman Pallet

Equilibrium's Bailsman Pallet is a Substrate module that manages users payment responsibility. Registered bailsmen receive fees payed by users and provide payment security for users positions

Re-exports

pub use eq_oracle;

Modules

rate
Module for Bailsman pallet interest rate calculations

Structs

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Module

Enums

Call
Dispatchable calls.

Error
Pallet's errors

Traits

BailsmanManager
Manages bailsman operation. OFFCHAIN ONLY

LtvChecker
LTV calculations and checks. LTV is a ratio of USD collateral value to USD debt value

Trait
Substrate pallet configuration trait

VolatilityGetter
Volatility, correlation and covariance storage adapter